+++ CHANGES.md
@@ -13,6 +13,7 @@
 * set default timeout to 12 hours, same as WP.org
 * fix for 3.9 setting theme update details to display:none;
 * fix for error when installing themes from WP.org repo
+* fix for incorrect plugin upgrade link in detail popup
 
 #### 2.6.3